+++ apps/dot-components-e2e/src/integration/NavigationRail/NavigationRail.spec.ts
@@ -1,0 +1,27 @@
+describe('dot-components: Navigation Rail component', () => {
+  before(() =>
+    cy.visit('/iframe.html?id=experimental-navigationrail--default')
+  );
+
+  it('should have a dot- prefix', () => {
+    cy.get('div').should('have.class', 'dot-navigation-rail');
+  });
+
+  describe('style decisions', () => {
+    it('container', () => {
+      cy.get('div.dot-navigation-rail')
+        .should('have.css', 'background-color', 'rgb(243, 245, 246)')
+        .and('have.css', 'width', '72px')
+        .and('have.css', 'padding-top', '8px')
+        .and('have.css', 'text-align', 'center');
+    });
+
+    it('selected item', () => {
+      cy.get('.rail-item-button.selected').should(
+        'have.css',
+        'background-color',
+        'rgb(255, 255, 255)'
+      );
+    });
+  });
+});